pythonPulse/requirements.txt
discord.py
python-dotenv
cohere
supabase
APScheduler
